refactor(schema): decouple MeasurementUnit and elevate currency to Journal
- Converted `MeasurementUnit` from Prisma Enum to TypeScript literal constant (`src/constants/enums.ts`), resolving generated import dependencies.
- Updated `schema.prisma` to move `originalAmount`, `currency`, and `exchangeRate` to the `Journal` model, ensuring base currency conversion is tracked at the root document level.
- Updated `Zod` validation schemas and `document_sync.repo.ts` to parse foreign currency details correctly during AI analysis.
