+++ src/app/api/v1/user/account_book/[account_book_id]/esg/route.ts
@@ -8,7 +8,7 @@ import { esgRepo } from "@/repositories/esg.repo";
 import { auditLogRepo } from "@/repositories/audit_log.repo";
 import { getIdentityFromDeWT } from "@/lib/auth/dewt";
 import { IEsgRecordFilterOptions } from "@/interfaces/data_filter_option";
-import { MeasurementUnit } from "@/generated";
+import { MeasurementUnit } from "@/constants/enums";
 
 /**
  * Info: (20260312 - Julian) 新增 ESG 紀錄
