+++ src/constants/enums.ts
@@ -1,0 +1,13 @@
+// Info: (20260514 - Tzuhan) Centralized constants to replace Prisma Enums
+
+export const MeasurementUnit = {
+  KWH: "KWH",
+  LITER: "LITER",
+  KG: "KG",
+  TONNE: "TONNE",
+  GALLON: "GALLON",
+  PIECE: "PIECE",
+} as const;
+
+export type MeasurementUnit =
+  (typeof MeasurementUnit)[keyof typeof MeasurementUnit];
